requirements.txt
faiss-cpu==1.8.0.post1
numpy==1.26.4
pypdf==4.2.0
sentence-transformers==2.7.0
